Refactor class property pages in to multiple modules
