+++ gaphor/UML/classes/classespropertypages.py
@@ -6,7 +6,6 @@ from gaphor import UML
 from gaphor.core import gettext, transactional
 from gaphor.core.format import format, parse
 from gaphor.diagram.propertypages import (
-    ComboModel,
     EditableTreeModel,
     PropertyPageBase,
     PropertyPages,
@@ -14,14 +13,10 @@ from gaphor.diagram.propertypages import (
     on_bool_cell_edited,
     on_text_cell_edited,
 )
-from gaphor.UML.classes.association import AssociationItem
 from gaphor.UML.classes.datatype import DataTypeItem
-from gaphor.UML.classes.dependency import DependencyItem
-from gaphor.UML.classes.enumeration import EnumerationItem
 from gaphor.UML.classes.interface import Folded, InterfaceItem
 from gaphor.UML.classes.klass import ClassItem
 from gaphor.UML.components.connector import ConnectorItem
-from gaphor.UML.profiles.stereotypepropertypages import stereotype_model
 
 log = logging.getLogger(__name__)
 
@@ -437,284 +432,3 @@ class OperationsPage(PropertyPageBase):
     def _on_show_operations_change(self, button, gparam):
         self.item.show_operations = button.get_active()
         self.item.request_update()
-
-
-@PropertyPages.register(EnumerationItem)
-class EnumerationPage(PropertyPageBase):
-    """An editor for enumeration literals for an enumeration."""
-
-    order = 20
-
-    def __init__(self, item):
-        super().__init__()
-        self.item = item
-        self.watcher = item.subject and item.subject.watcher()
-
-    def construct(self):
-        if not isinstance(self.item.subject, UML.Enumeration):
-            return
-
-        builder = new_builder("enumerations-editor")
-        page = builder.get_object("enumerations-editor")
-
-        show_enumerations = builder.get_object("show-enumerations")
-        show_enumerations.set_active(self.item.show_enumerations)
-
-        self.model = ClassEnumerationLiterals(self.item)
-
-        tree_view: Gtk.TreeView = builder.get_object("enumerations-list")
-        tree_view.set_model(self.model)
-
-        def handler(event):
-            enumeration = event.element
-            for row in self.model:
-                if row[-1] is enumeration:
-                    row[:] = [format(enumeration), enumeration]
-
-        self.watcher.watch("ownedLiteral.name", handler)
-
-        builder.connect_signals(
-            {
-                "show-enumerations-changed": (self._on_show_enumerations_change,),
-                "enumerations-name-edited": (on_text_cell_edited, self.model, 0),
-                "tree-view-destroy": (self.watcher.unsubscribe_all,),
-                "enumerations-keypress": (on_keypress_event,),
-            }
-        )
-        return page
-
-    @transactional
-    def _on_show_enumerations_change(self, button, gparam):
-        self.item.show_attributes = button.get_active()
-        self.item.request_update()
-
-
-PropertyPages.register(EnumerationItem)(AttributesPage)
-PropertyPages.register(EnumerationItem)(OperationsPage)
-
-
-@PropertyPages.register(DependencyItem)
-class DependencyPropertyPage(PropertyPageBase):
-    """Dependency item editor."""
-
-    order = 20
-
-    DEPENDENCY_TYPES = (
-        (gettext("Dependency"), UML.Dependency),
-        (gettext("Usage"), UML.Usage),
-        (gettext("Realization"), UML.Realization),
-        (gettext("Implementation"), UML.InterfaceRealization),
-    )
-
-    def __init__(self, item):
-        super().__init__()
-        self.item = item
-        self.watcher = self.item.watcher()
-        self.builder = new_builder(
-            "dependency-editor",
-            signals={
-                "dependency-type-changed": (self._on_dependency_type_change,),
-                "automatic-changed": (self._on_auto_dependency_change,),
-                "dependency-type-destroy": (self.watcher.unsubscribe_all,),
-            },
-        )
-
-    def construct(self):
-        dependency_combo = self.builder.get_object("dependency-combo")
-        model = ComboModel(self.DEPENDENCY_TYPES)
-        dependency_combo.set_model(model)
-
-        automatic = self.builder.get_object("automatic")
-        automatic.set_active(self.item.auto_dependency)
-
-        self.update()
-
-        self.watcher.watch("subject", self._on_subject_change)
-
-        return self.builder.get_object("dependency-editor")
-
-    def _on_subject_change(self, event):
-        self.update()
-
-    def update(self):
-        """Update dependency type combo box.
-
-        Disallow dependency type when dependency is established.
-        """
-        combo = self.builder.get_object("dependency-combo")
-        if combo.get_model():
-            item = self.item
-            index = combo.get_model().get_index(item.dependency_type)
-            combo.props.sensitive = not item.auto_dependency
-            combo.set_active(index)
-
-    @transactional
-    def _on_dependency_type_change(self, combo):
-        cls = combo.get_model().get_value(combo.get_active())
-        self.item.dependency_type = cls
-        subject = self.item.subject
-        if subject:
-            UML.model.swap_element(subject, cls)
-            self.item.request_update()
-
-    @transactional
-    def _on_auto_dependency_change(self, switch, gparam):
-        self.item.auto_dependency = switch.get_active()
-        self.update()
-
-
-def _dummy_handler(*args):
-    pass
-
-
-@PropertyPages.register(AssociationItem)
-class AssociationPropertyPage(PropertyPageBase):
-
-    NAVIGABILITY = (None, False, True)
-    AGGREGATION = ("none", "shared", "composite")
-
-    order = 20
-
-    def __init__(self, item):
-        self.item = item
-        self.subject = item.subject
-        self.watcher = item.subject and self.subject.watcher()
-        self.semaphore = 0
-
-    def handlers_end(self, end_name, end):
-        subject = end.subject
-
-        stereotypes = UML.model.get_stereotypes(subject)
-        if stereotypes:
-            model, toggle_handler, set_value_handler = stereotype_model(subject)
-            return model, {
-                f"{end_name}-toggle-stereotype": toggle_handler,
-                f"{end_name}-set-slot-value": set_value_handler,
-            }
-        else:
-            return None, {
-                f"{end_name}-toggle-stereotype": (_dummy_handler,),
-                f"{end_name}-set-slot-value": (_dummy_handler,),
-            }
-
-    def construct_end(self, builder, end_name, end, stereotypes_model):
-        subject = end.subject
-        title = builder.get_object(f"{end_name}-title")
-        if subject.type:
-            title.set_text(f"{end_name.title()} (: {subject.type.name})")
-
-        self.update_end_name(builder, end_name, subject)
-
-        navigation = builder.get_object(f"{end_name}-navigation")
-        navigation.set_active(self.NAVIGABILITY.index(subject.navigability))
-
-        aggregation = builder.get_object(f"{end_name}-aggregation")
-        aggregation.set_active(self.AGGREGATION.index(subject.aggregation))
-
-        if stereotypes_model:
-            stereotype_list = builder.get_object(f"{end_name}-stereotype-list")
-            stereotype_list.set_model(stereotypes_model)
-        else:
-            stereotype_frame = builder.get_object(f"{end_name}-stereotype-frame")
-            stereotype_frame.hide()
-
-    def update_end_name(self, builder, end_name, subject):
-        name = builder.get_object(f"{end_name}-name")
-        new_name = (
-            format(
-                subject,
-                visibility=True,
-                is_derived=True,
-                multiplicity=True,
-            )
-            or ""
-        )
-        if not (name.is_focus() or self.semaphore):
-            self.semaphore += 1
-            name.set_text(new_name)
-            self.semaphore -= 1
-        return name
-
-    def construct(self):
-        if not self.subject:
-            return None
-
-        head = self.item.head_end
-        tail = self.item.tail_end
-
-        head_model, head_signal_handlers = self.handlers_end("head", head)
-        tail_model, tail_signal_handlers = self.handlers_end("tail", tail)
-
-        builder = new_builder(
-            "association-editor",
-            signals={
-                "show-direction-changed": (self._on_show_direction_change,),
-                "invert-direction-changed": (self._on_invert_direction_change,),
-                "head-name-changed": (self._on_end_name_change, head),
-                "head-navigation-changed": (self._on_end_navigability_change, head),
-                "head-aggregation-changed": (self._on_end_aggregation_change, head),
-                "tail-name-changed": (self._on_end_name_change, tail),
-                "tail-navigation-changed": (self._on_end_navigability_change, tail),
-                "tail-aggregation-changed": (self._on_end_aggregation_change, tail),
-                "association-editor-destroy": (self.watcher.unsubscribe_all,),
-                **head_signal_handlers,
-                **tail_signal_handlers,
-            },
-        )
-
-        show_direction = builder.get_object("show-direction")
-        show_direction.set_active(self.item.show_direction)
-
-        self.construct_end(builder, "head", head, head_model)
-        self.construct_end(builder, "tail", tail, tail_model)
-
-        def name_handler(event):
-            end_name = "head" if event.element is head.subject else "tail"
-            self.update_end_name(builder, end_name, event.element)
-
-        def restore_nav_handler(event):
-            for end_name, end in (("head", head), ("tail", tail)):
-                combo = builder.get_object(f"{end_name}-navigation")
-                self._on_end_navigability_change(combo, end)
-
-        # Watch on association end:
-        self.watcher.watch("memberEnd[Property].name", name_handler).watch(
-            "memberEnd[Property].visibility", name_handler
-        ).watch("memberEnd[Property].lowerValue", name_handler).watch(
-            "memberEnd[Property].upperValue", name_handler
-        ).watch(
-            "memberEnd[Property].type",
-            restore_nav_handler,
-        )
-
-        return builder.get_object("association-editor")
-
-    @transactional
-    def _on_show_direction_change(self, button, gparam):
-        self.item.show_direction = button.get_active()
-
-    @transactional
-    def _on_invert_direction_change(self, button):
-        self.item.invert_direction()
-
-    @transactional
-    def _on_end_name_change(self, entry, end):
-        if not self.semaphore:
-            self.semaphore += 1
-            parse(end.subject, entry.get_text())
-            self.semaphore -= 1
-
-    @transactional
-    def _on_end_navigability_change(self, combo, end):
-        if end.subject and end.subject.opposite and end.subject.opposite.type:
-            UML.model.set_navigability(
-                end.subject.association,
-                end.subject,
-                self.NAVIGABILITY[combo.get_active()],
-            )
-            # Call this again, or non-navigability will not be displayed
-            self.item.update_ends()
-
-    @transactional
-    def _on_end_aggregation_change(self, combo, end):
-        end.subject.aggregation = self.AGGREGATION[combo.get_active()]
